refactor: move returnstyle to lib
