+++ common/src/asset.rs
@@ -538,3 +538,19 @@ mod tests {
         assert_eq!(deserialized, amount);
     }
 }
+
+#[derive(Clone, Debug)]
+#[near(serializers = [json])]
+pub enum ReturnStyle {
+    Nep141FtTransferCall,
+    Nep245MtTransferCall,
+}
+
+impl ReturnStyle {
+    pub fn serialize(&self, amount: FungibleAssetAmount<impl AssetClass>) -> serde_json::Value {
+        match self {
+            Self::Nep141FtTransferCall => serde_json::json!(amount),
+            Self::Nep245MtTransferCall => serde_json::json!([amount]),
+        }
+    }
+}
