+++ contract/market/src/lib.rs
@@ -114,25 +114,6 @@ mod impl_helper;
 mod impl_market_external;
 mod impl_token_receiver;
 
-#[derive(Clone, Debug)]
-#[near(serializers = [json])]
-pub enum ReturnStyle {
-    Nep141FtTransferCall,
-    Nep245MtTransferCall,
-}
-
-impl ReturnStyle {
-    pub fn serialize(
-        &self,
-        amount: templar_common::asset::FungibleAssetAmount<impl templar_common::asset::AssetClass>,
-    ) -> serde_json::Value {
-        match self {
-            Self::Nep141FtTransferCall => serde_json::json!(amount),
-            Self::Nep245MtTransferCall => serde_json::json!([amount]),
-        }
-    }
-}
-
 #[cfg(target_arch = "wasm32")]
 mod custom_getrandom {
     #![allow(clippy::no_mangle_with_rust_abi)]
